api/requirements.txt
fastapi
uvicorn[standard]
asyncpg
motor
pydantic
bcrypt
python-jose
python-dotenv
